refactor(app): isolate legacy home
